+++ evm/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [`v0.1.1`](https://github.com/ignite/apps/releases/tag/evm/v0.1.1)
 
-- []() Wire EVM mempool and correct start command.
+- [#235](https://github.com/ignite/apps/pull/235) Wire EVM mempool and correct start command.
 
 ## [`v0.1.0`](https://github.com/ignite/apps/releases/tag/evm/v0.1.0)